fc_circularizacao/requirements.txt
python-dotenv==1.0.1
pydantic==2.8.2
reportlab==4.2.2
PyPDF2==3.0.1
